+++ sites/m.dishen.com/components/mip-dishen-datepicker/example/index.html
@@ -10,13 +10,13 @@
   <body>
     <div>
       选择日期
-      <mip-dishen-datepicker data-date="2000-1-1" data-toid-date="date"  data-confirm="false" data-confirm="true" data-type="0">
+      <mip-dishen-datepicker data-date="2000-1-1" data-toid-date="date" data-type="0">
           <input type="hidden" name="date" id="date" value="0" on="change:MIP.setData({date:event.value})">
       </mip-dishen-datepicker>
     </div>
     <div>
       选择时间
-      <mip-dishen-datepicker data-hour="0" data-min="0" data-toid-hour="hour" data-toid-min="min" data-confirm="false" data-confirm="true" data-fixed="true">
+      <mip-dishen-datepicker data-hour="0" data-min="0" data-toid-hour="hour" data-toid-min="min" data-fixed="true" data-confirm="false">
           <input type="hidden" name="hour" id="hour" value="0" on="change:MIP.setData({hour:event.value})">
           <input type="hidden" name="min" id="min" value="0" on="change:MIP.setData({min:event.value})">
       </mip-dishen-datepicker>
