+++ sites/m.dishen.com/components/mip-dishen-datepicker/style/animation.less
@@ -1,0 +1,80 @@
+.mip-dishen-geardate {
+  .slideInUp {
+    animation: slideInUp 0.3s ease-in;
+    -webkit-animation: slideInUp 0.3s ease-in;
+    -ms-animation: slideInUp 0.3s ease-in;
+    -moz-animation: slideInUp 0.3s ease-in;
+  }
+
+  @-webkit-keyframes slideInUp {
+    0% {
+      -webkit-transform: translateY(100%);
+      transform: translateY(100%);
+      -moz-transform: translateY(100%);
+      -ms-transform: translateY(100%);
+      -o-transform: translateY(100%);
+    }
+
+    100% {
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+      -moz-transform: translateY(0);
+      -o-transform: translateY(0);
+      -ms-transform: translateY(0);
+    }
+  }
+
+  @keyframes slideInUp {
+    0% {
+      -webkit-transform: translateY(100%);
+      transform: translateY(100%);
+      -moz-transform: translateY(100%);
+      -ms-transform: translateY(100%);
+      -o-transform: translateY(100%);
+    }
+
+    100% {
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+      -moz-transform: translateY(0);
+      -o-transform: translateY(0);
+      -ms-transform: translateY(0);
+    }
+  }
+
+  @-moz-keyframes slideInUp {
+    0% {
+      -webkit-transform: translateY(100%);
+      transform: translateY(100%);
+      -moz-transform: translateY(100%);
+      -ms-transform: translateY(100%);
+      -o-transform: translateY(100%);
+    }
+
+    100% {
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+      -moz-transform: translateY(0);
+      -o-transform: translateY(0);
+      -ms-transform: translateY(0);
+    }
+  }
+
+  @-ms-keyframes slideInUp {
+    0% {
+      -webkit-transform: translateY(100%);
+      transform: translateY(100%);
+      -moz-transform: translateY(100%);
+      -ms-transform: translateY(100%);
+      -o-transform: translateY(100%);
+    }
+
+    100% {
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+      -moz-transform: translateY(0);
+      -o-transform: translateY(0);
+      -ms-transform: translateY(0);
+    }
+  }
+}
