+++ sites/m.dishen.com/components/mip-dishen-datepicker/style/roll.less
@@ -1,0 +1,71 @@
+
+.mip-dishen-geardate {
+  .date_roll,
+  .date_roll > div {
+    background-color: transparent;
+    overflow: hidden;
+  }
+
+  .date_roll {
+    position: relative;
+    width: 100%;
+    height: auto;
+    font-weight: 700;
+    -webkit-mask:
+      -webkit-gradient(
+        linear,
+        0 50%,
+        0 100%,
+        from(#debb47),
+        to(rgba(36, 142, 36, 0))
+      );
+    -webkit-mask:
+      -webkit-linear-gradient(
+        top,
+        #debb47 50%,
+        rgba(36, 142, 36, 0)
+      );
+  }
+
+  .date_roll > div {
+    font-size: 2em;
+    height: 6em;
+    float: left;
+    position: relative;
+    width: 33.33%;
+  }
+
+  .date_roll_col2 > div {
+    width: 50%;
+  }
+
+  .date_roll_col3 > div {
+    width: 33%;
+  }
+
+  .date_roll_col4 > div {
+    width: 25%;
+  }
+
+  .date_roll_col5 > div {
+    width: 20%;
+  }
+
+  .date_roll_mask {
+    padding: 0;
+    -webkit-mask:
+      -webkit-gradient(
+        linear,
+        0 40%,
+        0 0,
+        from(#debb47),
+        to(rgba(36, 142, 36, 0))
+      );
+    -webkit-mask:
+      -webkit-linear-gradient(
+        bottom,
+        #debb47 50%,
+        rgba(36, 142, 36, 0)
+      );
+  }
+}
