+++ sites/m.dishen.com/components/mip-dishen-xzpicker/astro-data.js
@@ -1,0 +1,72 @@
+export const astro = [{
+  index: '1',
+  name: '白羊座',
+  date: '3日21日-4月20日',
+  value: 'aries'
+},
+{
+  index: '2',
+  name: '金牛座',
+  date: '4月20日-5月20日',
+  value: 'taurus'
+},
+{
+  index: '3',
+  name: '双子座',
+  date: '5月21日-6月21日',
+  value: 'gemini'
+},
+{
+  index: '4',
+  name: '巨蟹座',
+  date: '6月22日-7月22日',
+  value: 'cancer'
+},
+{
+  index: '5',
+  name: '狮子座',
+  date: '7月23日-8月22日',
+  value: 'leo'
+},
+{
+  index: '6',
+  name: '处女座',
+  date: '8月23日-9月22日',
+  value: 'virgo'
+},
+{
+  index: '7',
+  name: '天秤座',
+  date: '9月23日-10月23日',
+  value: 'libra'
+},
+{
+  index: '8',
+  name: '天蝎座',
+  date: '10月24日-11月22日',
+  value: 'scorpio'
+},
+{
+  index: '9',
+  name: '射手座',
+  date: '11月23日-12月21日',
+  value: 'sagittarius'
+},
+{
+  index: '10',
+  name: '摩羯座',
+  date: '12月22日-1月19日',
+  value: 'capricorn'
+},
+{
+  index: '11',
+  name: '水瓶座',
+  date: '1月20日-2月18日',
+  value: 'aquarius'
+},
+{
+  index: '12',
+  name: '双鱼座',
+  date: '2月19日-3月20日',
+  value: 'pisces'
+}]
