+++ sites/m.dishen.com/components/mip-dishen-xzpicker/example/index.html
@@ -84,7 +84,8 @@
       <mip-dishen-xzpicker
         fetch-url="http://localhost:8099/pair"
         data-href="https://m.dishen.com/xz/pair/_val_.html"
-        >选择星座配对</mip-dishen-xzpicker>
+      >选择星座配对
+      </mip-dishen-xzpicker>
     </div>
     <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
     <script src="/mip-dishen-xzpicker/mip-dishen-xzpicker.js"></script>
