+++ sites/m.dishen.com/components/mip-dishen-xzpicker/mip-dishen-xzpicker.js
@@ -1,113 +1,43 @@
 import './index.less'
+import { astro } from './astro-data'
+import XzScroll from './XzScroll'
 // 星座选择控件
-function XzPicker (options) {
-  let defaults = {
-    index: 5, // 高亮位置
-    startX: 0, // 起始坐标x
-    startY: 0, // 起始坐标y
-    storeY: 0, // 保存y坐标
-    scrollY: !1,
-    astro: [{
-      index: '1',
-      name: '白羊座',
-      date: '3日21日-4月20日',
-      value: 'aries'
-    },
-    {
-      index: '2',
-      name: '金牛座',
-      date: '4月20日-5月20日',
-      value: 'taurus'
-    },
-    {
-      index: '3',
-      name: '双子座',
-      date: '5月21日-6月21日',
-      value: 'gemini'
-    },
-    {
-      index: '4',
-      name: '巨蟹座',
-      date: '6月22日-7月22日',
-      value: 'cancer'
-    },
-    {
-      index: '5',
-      name: '狮子座',
-      date: '7月23日-8月22日',
-      value: 'leo'
-    },
-    {
-      index: '6',
-      name: '处女座',
-      date: '8月23日-9月22日',
-      value: 'virgo'
-    },
-    {
-      index: '7',
-      name: '天秤座',
-      date: '9月23日-10月23日',
-      value: 'libra'
-    },
-    {
-      index: '8',
-      name: '天蝎座',
-      date: '10月24日-11月22日',
-      value: 'scorpio'
-    },
-    {
-      index: '9',
-      name: '射手座',
-      date: '11月23日-12月21日',
-      value: 'sagittarius'
-    },
-    {
-      index: '10',
-      name: '摩羯座',
-      date: '12月22日-1月19日',
-      value: 'capricorn'
-    },
-    {
-      index: '11',
-      name: '水瓶座',
-      date: '1月20日-2月18日',
-      value: 'aquarius'
-    },
-    {
-      index: '12',
-      name: '双鱼座',
-      date: '2月19日-3月20日',
-      value: 'pisces'
-    }
-    ],
-    callback: function (data) {
-      {
-        let AJAXurl = options.fetchUrl + '?girlCaturl=' + data.girl.index + '&boyCaturl=' + data.boy.index + ''
-        fetch(AJAXurl).then(function (res) {
-          return res.json()
-        }).then(function (data) {
-          if (data.code === 200) {
-            if (options.href) {
-              let href = options.href.replace('_val_', '' + data.pair_id)
-              window.location.href = href
+class XzPicker {
+  constructor (options) {
+    let defaults = {
+      index: 5, // 高亮位置
+      startX: 0, // 起始坐标x
+      startY: 0, // 起始坐标y
+      storeY: 0, // 保存y坐标
+      scrollY: !1,
+      astro: astro,
+      callback: function (data) {
+        {
+          let AJAXurl = options.fetchUrl + '?girlCaturl=' + data.girl.index + '&boyCaturl=' + data.boy.index + ''
+          fetch(AJAXurl).then(function (res) {
+            return res.json()
+          }).then(function (data) {
+            if (data.code === 200) {
+              if (options.href) {
+                let href = options.href.replace('_val_', '' + data.pair_id)
+                window.location.href = href
+              }
+            } else {
+              console.log(data.msg)
             }
-          } else {
-            console.log(data.msg)
-          }
-        }).catch(function (ex) {
-          // 捕获 parse的异常
-          console.log('parsing failed', ex)
-        })
+          }).catch(function (ex) {
+            // 捕获 parse的异常
+            console.log('parsing failed', ex)
+          })
+        }
       }
     }
+    this.params = defaults
   }
-  this.params = defaults
-}
-XzPicker.prototype = {
-  createHTML: function () {
+  createHTML () {
     let listStr = ''
     for (let i = 0; i < this.params.astro.length; i++) {
-      listStr += ' <li><span class="pic xz-icon_' + (i + 1) + '"></span><div class="tit">' + this.params.astro[i].name + '</div></li>'
+      listStr += ' <li><span class="pic astro-icon icon' + (i + 1) + '"></span><div class="tit">' + this.params.astro[i].name + '</div></li>'
     }
     let html =
             '<div class="xz-picker-inner">' +
@@ -137,12 +67,12 @@ XzPicker.prototype = {
             '</div>' +
             '<div class="xz-selector-overlay"></div>'
     this.pickerEl = document.createElement('div')
-    this.pickerEl.className = 'xz-picker on'
+    this.pickerEl.className = 'mip-dishen-xzpicker on'
     this.pickerEl.id = 'xzPicker'
     this.pickerEl.innerHTML = html
     document.body.appendChild(this.pickerEl)
-  },
-  events: function () {
+  }
+  events () {
     let _self = this
     let pickerList = _self.pickerEl.querySelectorAll('.picker-list')
     new XzScroll().init(pickerList[0])
@@ -163,8 +93,8 @@ XzPicker.prototype = {
         boy: boyData
       })
     })
-  },
-  init: function (el) {
+  }
+  init (el) {
     this.trigger = el
     let that = this
     this.trigger.addEventListener('click', function (e) {
@@ -178,158 +108,7 @@ XzPicker.prototype = {
     })
   }
 }
-// 星座选择滚动scroll
-function XzScroll (options) {
-  let defaults = {
-    index: 5,
-    startX: 0, // 记录起始位置
-    startY: 0, // 记录起始位置
-    scrollY: 0,
-    offsetY: 0, // 当前滑动偏移量
-    recordY: 0, // 记录滑动后conBox偏移量
-    timestamp: 0,
-    speed: 400
-  }
-  this.params = defaults
-}
-XzScroll.prototype = {
-  init: function (el) {
-    let _self = this
-    _self.dom = el
-    _self.params.list = el.querySelector('ul')
-    _self.params.items = el.querySelectorAll('li')
-    _self.params.itemHeight = _self.params.items[0].offsetHeight
-    _self.params.scrollHeight = _self.params.list.offsetHeight - el.offsetHeight // 可滚动的高度
-    if (_self.params.index) {
-      _self.params.offsetY = -_self.params.index * _self.params.itemHeight
-      _self.params.recordY = _self.params.offsetY
-      _self.params.list.style['transform'] = 'translate(0,' + _self.params.offsetY + 'px)'
-      el.setAttribute('data-index', _self.params.index)
-    }
-    this.bindEvents()
-  },
-  bindEvents: function () {
-    let _self = this
 
-    function correctOffsetY (_offsetY) {
-      // 超出容器左边界
-      if (-_offsetY <= 0) {
-        _offsetY = Math.pow(_offsetY, 0.8)
-      } else if (-_offsetY >= _self.params.scrollHeight) { // 超出容器右边界
-        _offsetY = -_self.params.scrollHeight - Math.pow(-_offsetY - _self.params.scrollHeight, 0.8)
-      }
-      return _offsetY
-    }
-    _self.params.list.addEventListener('touchstart', function (e) {
-      _self.params.startX = e.touches[0].pageX
-      _self.params.startY = e.touches[0].pageY
-      _self.params.scrollY = undefined
-      _self.params.timestamp = new Date().getTime()
-    }, false)
-    _self.params.list.addEventListener('touchmove', function (e) {
-      let distX = e.touches[0].pageX - _self.params.startX
-      let distY = e.touches[0].pageY - _self.params.startY
-      if (typeof _self.params.scrollY === 'undefined') {
-        _self.params.scrollY = Math.abs(distX) < Math.abs(distY)
-      }
-      if (_self.params.scrollY) {
-        e.preventDefault()
-        // 当前滑动偏移
-        _self.params.offsetY = distY + _self.params.recordY
-        // 偏移限定纠正
-        _self.params.offsetY = correctOffsetY(_self.params.offsetY)
-        _self.params.list.style['transform'] = 'translate(0,' + _self.params.offsetY + 'px)'
-      }
-    }, false)
-    // 释放触摸，回滚
-    _self.params.list.addEventListener('touchend', function (e) {
-      let s = e.changedTouches[0].pageY - _self.params.startY
-      let t = new Date().getTime() - _self.params.timestamp
-      let v = Math.abs(s) / t
-      if (v > 0.3) {
-        if (s > 0) {
-          _self.params.offsetY -= v * _self.params.offsetY * 1.2
-        } else {
-          _self.params.offsetY += v * _self.params.offsetY * 1.2
-        }
-      }
-      if (_self.params.offsetY > 0) {
-        _self.params.index = 0
-        _self.params.recordY = 0
-        _self.params.offsetY = 0
-        doAnimation()
-        // _self.params.list.style["transform"] = 'translate(0,0px)';
-      } else if (-_self.params.offsetY > _self.params.scrollHeight) {
-        _self.params.index = _self.params.items.length - 1
-        _self.params.recordY = -_self.params.scrollHeight
-        _self.params.offsetY = -_self.params.scrollHeight
-        // _self.params.list.style["transform"] = 'translate(0,-' + _self.params.scrollHeight + 'px)';
-        doAnimation()
-      } else {
-        let count = parseInt((_self.params.offsetY - _self.params.itemHeight / 2) / _self.params.itemHeight)
-        _self.params.index = Math.abs(count)
-        _self.params.offsetY = count * _self.params.itemHeight
-        _self.params.recordY = _self.params.offsetY
-        // _self.params.list.style["transform"] = 'translate(0,' + _self.params.offsetY + 'px)';
-        doAnimation()
-      }
-      _self.dom.setAttribute('data-index', _self.params.index)
-    }, false)
-    // 缓动动画
-    function doAnimation () {
-      let curY = parseInt(_self.params.list.style['transform'].split(',')[1].replace('px)', ''))
-      runAnimation(curY, _self.params.offsetY, 300, function (value) {
-        _self.params.list.style['transform'] = 'translate(0,' + value + 'px)'
-      })
-    }
-  }
-}
-// 对运动方法进行封装
-function runAnimation (from, to, duration, callback) {
-  let options = {
-    duration: duration || 300,
-    callback: callback || null
-  }
-  // requestAnimationFrame的兼容处理
-  if (!window.requestAnimationFrame) {
-    window.requestAnimationFrame2 = function (fn) {
-      setTimeout(fn, 17)
-    }
-  } else {
-    window.requestAnimationFrame2 = window.requestAnimationFrame
-  }
-  // 算法需要的几个变量
-  let start = 0
-  // during根据设置的总时间计算
-  let during = Math.ceil(options.duration / 17)
-  /*
-     * t: current time（当前时间）；
-     * b: beginning value（初始值）；
-     * c: change in value（变化量）；
-     * d: duration（持续时间）。
-     * you can visit 'http://easings.net/zh-cn' to get effect
-     */
-  let fnGetValue = function (t, b, c, d) {
-    return c * ((t = t / d - 1) * t * t + 1) + b
-  }
-  // 运动
-  let step = function () {
-    // 当前的运动位置
-    let value = fnGetValue(start, from, to - from, during)
-    // 时间递增
-    start++
-    // 如果还没有运动到位，继续
-    if (start <= during) {
-      options.callback(value)
-      requestAnimationFrame(step)
-    } else {
-      // 动画结束，这里可以插入回调...
-      options.callback(to, true)
-    }
-  }
-  // 开始执行动画
-  step()
-}
 export default class MIPXzpicker extends MIP.CustomElement {
   build () {
     let el = this.element
